+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add config for unplaceable items
 - Add tag 'plonk:unplaceable' for unplaceable items
 - Add safer tile entity getting for non-essential methods
+- Fix crash when running on dedicated server
 
 ## [7.0.4] - 2021-01-19